requirements.txt
# Core ML dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data collection and processing
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter for experimentation
jupyter>=1.0.0
jupyterlab>=4.0.0

# Optional: Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0

# Optional: Enhanced text processing (if needed)
# nltk>=3.8
# spacy>=3.6.0
# transformers>=4.30.0

# Optional: Deployment (if needed)
# flask>=2.3.0
# fastapi>=0.100.0
# uvicorn>=0.23.0